+++ core/src/main/java/de/bitbrain/braingdx/graphics/pipeline/RenderPipe.java
@@ -31,4 +31,6 @@ public interface RenderPipe {
    void addEffects(PostProcessorEffect... effects);
 
    void setEffects(PostProcessorEffect[] effects);
+
+   boolean hasEffects();
 }